+++ api/python/quilt3/packages.py
@@ -41,9 +41,11 @@ from .data_transfer import (
 from .exceptions import PackageException
 from .formats import CompressionRegistry, FormatRegistry
 from .telemetry import ApiTelemetry
-from .util import CACHE_PATH, DISABLE_TQDM, PACKAGE_UPDATE_POLICY
-from .util import TEMPFILE_DIR_PATH as APP_DIR_TEMPFILE_DIR
 from .util import (
+    CACHE_PATH,
+    DISABLE_TQDM,
+    PACKAGE_UPDATE_POLICY,
+    TEMPFILE_DIR_PATH as APP_DIR_TEMPFILE_DIR,
     PhysicalKey,
     QuiltConflictException,
     QuiltException,
@@ -400,7 +402,7 @@ class ManifestJSONDecoder(json.JSONDecoder):
     """
 
     def __init__(self, *args, **kwargs):
-        @functools.lru_cache(maxsize=None)
+        @functools.cache
         def memoize_key(s):
             return s
 
@@ -450,7 +452,7 @@ class Package:
         # traverse the tree of package directories and entries to get the list of
         # display objects. candidates is a deque of shape
         # ((logical_key, Package | PackageEntry), [list of parent key])
-        candidates = deque(([x, []] for x in self._children.items()))
+        candidates = deque([x, []] for x in self._children.items())
         results_dict = {}
         results_total = 0
         more_objects_than_lines = False
@@ -1724,7 +1726,7 @@ class Package:
             elif entry != other_entry:
                 modified.append(lk)
 
-        added = list(sorted(other_entries))
+        added = sorted(other_entries)
 
         return added, modified, deleted