+++ catalog/app/containers/Bucket/Queries/Athena/Database.spec.tsx
@@ -1,5 +1,5 @@
 import * as React from 'react'
-import renderer from 'react-test-renderer'
+import { render } from '@testing-library/react'
 
 import WithGlobalDialogs from 'utils/GlobalDialogs'
 
@@ -49,16 +49,16 @@ describe('containers/Bucket/Queries/Athena/Database', () => {
   afterAll(() => {})
 
   it('should render skeletons', () => {
-    const tree = renderer.create(
+    const { container } = render(
       <Provider value={emptyState}>
         <Database />
       </Provider>,
     )
-    expect(tree).toMatchSnapshot()
+    expect(container.firstChild).toMatchSnapshot()
   })
 
   it('should render selected values', () => {
-    const tree = renderer.create(
+    const { container } = render(
       <Provider
         value={{
           ...emptyState,
@@ -71,11 +71,11 @@ describe('containers/Bucket/Queries/Athena/Database', () => {
         <Database />
       </Provider>,
     )
-    expect(tree).toMatchSnapshot()
+    expect(container.firstChild).toMatchSnapshot()
   })
 
   it('should show no value (zero-width space) if selected no value', () => {
-    const tree = renderer.create(
+    const { container } = render(
       <Provider
         value={{
           ...emptyState,
@@ -88,11 +88,11 @@ describe('containers/Bucket/Queries/Athena/Database', () => {
         <Database />
       </Provider>,
     )
-    expect(tree).toMatchSnapshot()
+    expect(container.firstChild).toMatchSnapshot()
   })
 
   it('should disable selection if no spare values', () => {
-    const tree = renderer.create(
+    const { container } = render(
       <Provider
         value={{
           ...emptyState,
@@ -105,11 +105,11 @@ describe('containers/Bucket/Queries/Athena/Database', () => {
         <Database />
       </Provider>,
     )
-    expect(tree).toMatchSnapshot()
+    expect(container.firstChild).toMatchSnapshot()
   })
 
   it('should show error when values failed', () => {
-    const tree = renderer.create(
+    const { container } = render(
       <WithGlobalDialogs>
         <Provider
           value={{
@@ -122,11 +122,11 @@ describe('containers/Bucket/Queries/Athena/Database', () => {
         </Provider>
       </WithGlobalDialogs>,
     )
-    expect(tree).toMatchSnapshot()
+    expect(container.firstChild).toMatchSnapshot()
   })
 
   it('should show error when data failed', () => {
-    const tree = renderer.create(
+    const { container } = render(
       <WithGlobalDialogs>
         <Provider
           value={{
@@ -139,6 +139,6 @@ describe('containers/Bucket/Queries/Athena/Database', () => {
         </Provider>
       </WithGlobalDialogs>,
     )
-    expect(tree).toMatchSnapshot()
+    expect(container.firstChild).toMatchSnapshot()
   })
 })